Rename to sync
