+++ src/modules/sync/components/sync-connected-sites.tsx
@@ -23,9 +23,9 @@ import { useSyncStatesProgressInfo } from 'src/hooks/use-sync-states-progress-in
 import { cx } from 'src/lib/cx';
 import { getIpcApi } from 'src/lib/get-ipc-api';
 import { getLocalizedLink } from 'src/lib/get-localized-link';
-import { OpenSitesSyncSelector } from 'src/modules/content-tab-sync';
-import { SyncDialog } from 'src/modules/content-tab-sync/components/sync-dialog';
-import { convertTreeToOptionsToSync } from 'src/modules/content-tab-sync/convert-tree-to-options-to-sync';
+import { OpenSitesSyncSelector } from 'src/modules/sync';
+import { SyncDialog } from 'src/modules/sync/components/sync-dialog';
+import { convertTreeToOptionsToSync } from 'src/modules/sync/convert-tree-to-options-to-sync';
 import { useI18nLocale } from 'src/stores';
 import type { SyncSite } from 'src/hooks/use-fetch-wpcom-sites/types';
