+++ src/modules/sync/components/sync-sites-modal-selector.tsx
@@ -13,7 +13,7 @@ import { useOffline } from 'src/hooks/use-offline';
 import { cx } from 'src/lib/cx';
 import { getIpcApi } from 'src/lib/get-ipc-api';
 import { getLocalizedLink } from 'src/lib/get-localized-link';
-import { CreateButton } from 'src/modules/content-tab-sync/components/create-button';
+import { CreateButton } from 'src/modules/sync/components/create-button';
 import { useI18nLocale } from 'src/stores';
 import type { SyncSite } from 'src/hooks/use-fetch-wpcom-sites/types';
