+++ src/modules/sync/images/sync-tab-image.tsx
